requirements.txt
opencv-python
ultralytics
Flask
numpy
pyyaml
torch
torchvision
torchaudio